chore(slice3 #52): consolidate apps/frontend/src/lib/api.ts into lib/api/
Split legacy single-file fetch wrappers into domain modules under lib/api/:
- auth.ts        — UnauthenticatedError, fetchMe, mockLogin, logout
- permissions.ts — fetchPermissionCheck, createPermissionRequest, fetchPermissionRequestsMine
- managed-systems.ts — MS DTO + 4 CRUD funcs
- analytics-areas.ts — AA DTO + 4 CRUD funcs

Drop duplicate ApiError / ApiErrorEnvelope (legacy had its own; now imports
from ./types which re-exports @fops/shared ErrorEnvelope as single source).

Update 10 consumers (8 source + 2 test) from '../lib/api.js' → '../lib/api'.

Internal fetch behavior preserved verbatim (credentials: 'same-origin',
manual envelope handling) — apiClient rewrite deferred as separate follow-up
to limit blast radius. #19 onward can pick the new modular surface.

Closes #52
