+++ apps/frontend/src/lib/api/index.ts
@@ -2,3 +2,7 @@ export * from './types';
 export * from './errorMapper';
 export * from './useIdempotencyKey';
 export * from './client';
+export * from './auth';
+export * from './permissions';
+export * from './managed-systems';
+export * from './analytics-areas';
