+++ apps/frontend/src/routes/-index.test.tsx
@@ -13,7 +13,7 @@ import {
 } from '@tanstack/react-router';
 import { render, screen, waitFor } from '@testing-library/react';
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
-import { UnauthenticatedError, fetchMe } from '../lib/api.js';
+import { UnauthenticatedError, fetchMe } from '../lib/api';
 import { HomePage } from './index';
 import { LoginPage } from './login';
